Complex analysis

Regions of the complex plane

A number of domains relate to the complex plane:

ComplexUnitCircle

ComplexUnitDisk

For a more specific package dealing with complex regions, see ComplexRegions.jl.

Identification with the plane

Any domain in $ℝ^2$ can be identified with a region of the complex plane. The relevant maps to use are FunctionMaps.VectorToComplex and FunctionMaps.ComplexToVector.

DomainSets.FunctionMaps.VectorToComplex — Type

VectorToComplex()
VectorToComplex{T}()

Map a length 2 vector $[a;b]$ to the complex number $a+bi$.

See also: ComplexToVector.

DomainSets.FunctionMaps.ComplexToVector — Type

ComplexToVector()
ComplexToVector{T}()

Map a complex number $a+bi$ to the length 2 vector $[a; b]$.

See also: VectorToComplex.